Initial commit.
Well, not really the first commit -- I originally wrote this as a small
utility for myself, with all settings (including Plex URL + token) in
the .py file.

The original implementation simply stored an episode GUID and its
summary in a .json file, and so was too limited when I wanted to
ignore certain shows.

After rewriting the program to use SQLite and a configuration file,
I re-started the Git repository, with this commit.