Add --config-dir and improve automatic detection
By default, the script looks for the config file in the same directory
as the .py file, and if not found there, also the in the parent
directory.

If --config-dir is specified, it only looks in that directory.

The database is stored in the same directory by default, but that can
be overridden in the configuration file.