Major rewrite -- remove database entirely
After realizing how easy it is to tell Plex to re-download the data,
I figured that it's not worth the effort to store the hidden summaries
in a local database. This simplifies the code a fair bit, and also
makes the script easier to use.

The only downside I can think of is for people who frequently edit the
summaries and want their edits to remain after using this script.
Sorry to the extremely few of you who that description matches.