Wait for Plex to finish prior to continuing
--restore-all can miss one or two summaries and leave them hidden,
and presumably this can also happen during normal usage.

It seems that either the Python Plex API or the actual Plex server API
either drops some requests, or processes them out of order.

If a refresh requests runs prior to unlocking the item, it would not
have any effect, and leave the summary hidden.
Therefore, with this commit, the script first unlocks everything, waits
for all unlocks to finish, and only then refreshes.

It also waits for Plex to finish all script-related operations before exiting.