Fix restore failure issue
The root cause is that the Plex metadata downloader sometimes fails, and
it doesn't retry, so summaries that it fails to download simply won't
appear. Trying again, even seconds later, usually works.

As such, the only solution I can see is to wait and then verify that
Plex has successfully updated the summaries, which this commit
implements.

This and the previous commit add much more complexity than I'd like, but
leaving restores to random chance doesn't seem okay.