Rewrite main logic.
The old code was really only intended to hide summaries, not titles or
thumbnails, so adding support for titles/thumbnails was a hack.

This commit changes the main logic to calculate a set of Actions that
transform the old state into the wanted state, instead of the old hack
of separate stages for restore / hide, which doesn't work very well when
you want to hide some fields, and restore others, on the same item (due
to config changes).