Rework thumbnail hiding entirely
Unfortunately, the previous solution was much nicer, but it didn't work
properly, due to weird Plex behavior: if we set the "thumb" field to ""
and lock it, Plex will still re-download the thumbnail on the next
refresh.
"summary" and "title" keep their value, as expected, however.

One way to solve this is to store the thumbnails next to each episode
(cluttery), and another is to upload them to Plex, which is what this
commit does.
As before, it uses the season poster (if available), otherwise the
show's.